Rename `src/context.ts` to `src/SafeProvider.ts`
