+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from '@/hooks/index.js'
 export * from '@/types/index.js'
-export { SafeProvider, SafeContext } from './context.js'
+export { SafeProvider, SafeContext } from './SafeProvider.js'
 export { createConfig } from './createConfig.js'
